# web/testdata/openapi.golden.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Golden API",
    "description": "Auto-generated API documentation for Golden API",
    "version": "1.0.0",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "API Team",
      "email": "api@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "paths": {
    "/health": {
      "get": {
        "operationId": "listHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Health check",
        "tags": [
          "System"
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "listUser",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "example": 50,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "admin",
                "user"
              ],
              "type": "string"
            }
          },
          {
            "description": "Page number (1-based)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "example": 1,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Sort expression, e.g. -created_at",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "example": "-created_at",
              "type": "string"
            }
          },
          {
            "description": "Opaque cursor for cursor-based pagination",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "example": "eyJpZCI6IjEyMyJ9",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/goldenUser"
                      },
                      "type": "array"
                    },
                    "next_cursor": {
                      "example": "eyJpZCI6IjEyMyJ9",
                      "nullable": true,
                      "type": "string"
                    },
                    "total": {
                      "example": 100,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "items",
                    "total"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "error": "Internal server error"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": null
          }
        ],
        "summary": "List all users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/goldenCreateUserBody"
              }
            }
          },
          "description": "Request payload",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/goldenUser"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create new user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "patch": {
        "operationId": "patchUserById",
        "parameters": [
          {
            "description": "The id identifier",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/goldenPatchUserBody"
                  },
                  {
                    "anyOf": [
                      {
                        "required": [
                          "email"
                        ]
                      },
                      {
                        "required": [
                          "role"
                        ]
                      }
                    ]
                  }
                ]
              }
            }
          },
          "description": "Request payload",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/goldenUser"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [],
        "summary": "Update user",
        "tags": [
          "Users"
        ]
      }
    }
  },
  "security": [
    {
      "BearerAuth": null
    }
  ],
  "tags": [
    {
      "name": "System",
      "description": "System endpoints"
    },
    {
      "name": "Users",
      "description": "User management"
    }
  ],
  "components": {
    "schemas": {
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "goldenCreateUserBody": {
        "example": {
          "email": "example string",
          "role": "example string"
        },
        "properties": {
          "email": {
            "format": "email",
            "type": "string"
          },
          "role": {
            "default": "user",
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          }
        },
        "required": [
          "email",
          "role"
        ],
        "type": "object"
      },
      "goldenPatchUserBody": {
        "example": {
          "email": "example string",
          "role": "example string"
        },
        "properties": {
          "email": {
            "nullable": true,
            "type": "string"
          },
          "role": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "goldenUser": {
        "example": {
          "email": "user@example.com",
          "id": "usr_123",
          "role": "example string"
        },
        "properties": {
          "email": {
            "example": "user@example.com",
            "type": "string"
          },
          "id": {
            "example": "usr_123",
            "type": "string"
          },
          "role": {
            "enum": [
              "admin",
              "user"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  }
}